+++ examples/load-obj-file.js
@@ -22,7 +22,18 @@ export var loadObjFile = function() {
 
     function onGround(m) {
       loadObjFile.addMeshObj({
-        material: {type: 'standard'},
+        material: {type: 'mirror'},
+        envMapParams: {
+          baseColorMix: 0.5,                // CLEAR SKY
+          mirrorTint: [0.9, 0.95, 1.0],     // Slight cool tint
+          reflectivity: 0.55,               // 25% reflection blend
+          illuminateColor: [0.2, 0.2, 0.2], // Soft cyan
+          illuminateStrength: 0.1,          // Gentle rim
+          illuminatePulse: 0.01,            // No pulse (static)
+          fresnelPower: 1.0,                // Medium-sharp edge
+          envLodBias: 1.5,
+          usePlanarReflection: false,       // ✅ Env map mode
+        },
         position: {x: 0, y: -5, z: -10},
         rotation: {x: 0, y: 0, z: 0},
         rotationSpeed: {x: 0, y: 0, z: 0},
@@ -39,21 +50,20 @@ export var loadObjFile = function() {
 
     async function onLoadObj(m) {
       loadObjFile.addMeshObj({
-        material: {type: 'standard'},
+        material: {type: 'mirror'},
         position: {x: 0, y: -1, z: -20},
         rotation: {x: 0, y: 0, z: 0},
         scale: [100, 100, 100],
         rotationSpeed: {x: 0, y: 0, z: 0},
-        // texturesPaths: ['./res/textures/cube-g1.webp'],
         texturesPaths: ['./res/textures/cube-g1.webp', './res/textures/env-maps/sky1_lod_mid.webp'],
         envMapParams: {
           baseColorMix: 0.0,                // CLEAR SKY
           mirrorTint: [0.9, 0.95, 1.0],     // Slight cool tint
-          reflectivity: 0.25,               // 25% reflection blend
-          illuminateColor: [0.3, 0.7, 1.0], // Soft cyan
+          reflectivity: 0.95,               // 25% reflection blend
+          illuminateColor: [0.2, 0.2, 0.2], // Soft cyan
           illuminateStrength: 0.1,          // Gentle rim
           illuminatePulse: 0.01,            // No pulse (static)
-          fresnelPower: 2.0,                // Medium-sharp edge
+          fresnelPower: 1.0,                // Medium-sharp edge
           envLodBias: 1.5,
           usePlanarReflection: false,       // ✅ Env map mode
         },
@@ -65,7 +75,7 @@ export var loadObjFile = function() {
         }
       });
 
-      loadObjFile.addMeshObj({
+      let MYCUBE = loadObjFile.addMeshObj({
         material: {type: 'standard'},
         position: {x: 0, y: 3, z: -10},
         rotation: {x: 0, y: 0, z: 0},
@@ -95,56 +105,40 @@ export var loadObjFile = function() {
         }
       })
 
-      // var glbFile11 = await fetch("res/meshes/glb/woman1.glb").then(res => res.arrayBuffer().then(buf => uploadGLBModel(buf, loadObjFile.device)));
-      //    loadObjFile.addGlbObjInctance({
-      //      material: {type: 'mirror', useTextureFromGlb: true},
-      //      envMapParams: {
-      //        baseColorMix: 0.75,
-      //        mirrorTint: [0.9, 0.5, 1.0],    // Slight cool tint
-      //        reflectivity: 0.5,               // 25% reflection blend
-      //        illuminateColor: [0.3, 0.7, 1.0], // Soft cyan
-      //        illuminateStrength: 0.1,          // Gentle rim
-      //        illuminatePulse: 0.001,             // No pulse (static)
-      //        fresnelPower: 5.0,                // Medium-sharp edge
-      //        envLodBias: 2.5,
-      //        usePlanarReflection: false,  // ✅ Env map mode
-      //      },
-      //      useScale: true,
-      //      scale: [20, 20, 20],
-      //      position: {x: 0, y: -4, z: -20},
-      //      name: 'woman1',
-      //      texturesPaths: ['./res/meshes/glb/textures/mutant_origin.webp', './res/textures/env-maps/sky1.webp'],
-      //    }, null, glbFile11);
 
 
-
-      loadObjFile.lightContainer[0].intensity = 10;
-
+      loadObjFile.lightContainer[0].intensity = 1;
       loadObjFile.activateBloomEffect();
-      loadObjFile.lightContainer[0].behavior.setOsc0(-2, 2, 0.001)
+      loadObjFile.lightContainer[0].behavior.setOsc0(-2, 2, 0.01)
       loadObjFile.lightContainer[0].behavior.value_ = -1;
       loadObjFile.lightContainer[0].updater.push((light) => {
-        light.position[0] = light.behavior.setPath0()
-        light.target[0] = light.behavior.setPath0()
+        light.setTargetX(light.behavior.setPath0());
+        light.setPosX(light.behavior.setPath0());
       })
 
-      loadObjFile.lightContainer[0].position = [0, 17, -10];
-      loadObjFile.lightContainer[0].target = [0, 0, -10];
+      loadObjFile.lightContainer[0].setPosition([0, 15, -10]);
+      loadObjFile.lightContainer[0].setTarget([0, 0, -10]);
 
-      var TEST = loadObjFile.getSceneObjectByName('cube2');
       setTimeout(() => {
-        let cube1 = app.getSceneObjectByName('cube1')
-        // cube1.effects.flameEffect.intensity = 100;
-        // cube1.effects.flameEffect.morphTo("pyramid", 8)
+        MYCUBE.effects.flameEmitter.setIntensity(100);
+        MYCUBE.effects.flameEmitter.recreateVertexDataCrazzy(4);
+        MYCUBE.setAmbient(10, 1, 0);
         app.cameras.WASD.yaw = -0.03;
         app.cameras.WASD.pitch = -0.49;
         app.cameras.WASD.position[2] = 0;
-        app.cameras.WASD.position[1] = 5;
-      }, 800);
+        app.cameras.WASD.position[1] = 10;
+        app.cameras.WASD._dirtyAngle = true;
+
+      }, 400);
     }
 
     loadObjFile.canvas.addEventListener("ray.hit.event", (e) => {
       console.log('ray.hit.event detected');
+
+      if(e.detail.hitObject.name.startWith('cube')) {
+        e.detail.hitObject.effects.flameEmitter.recreateVertexDataCrazzy(4)
+      }
+
     });
 
   })